fix: move export_library_v2_to_zip to backup file as well
